feat: Track enrollment in courses
Courses shouldn't have more than 20 students. Since this is a number we'll probably reference frequently, we'll want to keep a separate tracking variable in each course. We're denormalizing our fake db, which is appropriate at times - in this case it's easier to keep track of data two places than to recalculate it repeatedly.